Extract logging into its own module
